Make an options struct for dispenso::parallel_for, add a way to limit threads in parallel_for
Summary: This options struct only grows the number of options by 1, but it will make any further changes easier to push out, because it will likely allow us to maintain this function signature going forward.  Also added an option for limiting the number of threads to use in the parallel_for beyond the inherent limits imposed by the utilized thread pool.

Differential Revision: D22279287

fbshipit-source-id: 2d8e84a16e52f13efa832896f31a8ce718db7d44