Add benchmark and test for when_all-based tree building
Summary:
Check performance from when_all by adding a benchmark to future_benchmark.  Add a similar test at the same time.  Benchmark shows a pretty bad slowdown for this technique.  Need to consider if there's a smarter way to write this.  One problem is that we don't have unwrapping constructors, and so we wind up with nesting of futures which is not performant.  One other potential issue is use of shared_ptr, which results in memory allocation, not a good thing for parallel code.

So far, this is a bit dissappointing, but at least we have a working basis for making a better version.

Differential Revision: D22038918

fbshipit-source-id: 01d7ba0a818521775814c18fb84bb6c5e5aadd99