Avoid potential unsequenced condition
Summary: A bug was found when building with gcc that caused a crash.  This was due to using a reference to an object in a shared_ptr and moving the shared_ptr in an unsequenced way.  On some compilers the reference is taken first, on others the reference is taken after the std::move().  After the fix, the reference is guaranteed to be taken prior to the move.

Reviewed By: antonsoch

Differential Revision: D22603884

fbshipit-source-id: 168e3f65d22f46277ebacffe99e833794b833249