Fix bug in dispenso SmallBufferAllocator
Summary:
This bug happened because there is a case in the test where *only* deallocation happens within some threads.  In that case, the registration of thread buffer return to central storage never occurred, and we "leaked" memory (okay, leak checkers wouldn't find this because we still have a handle, but we could essentially never use the memory again).

The ideal fix would be to register using a static thread_local member, but this appears not to work as intended against more than one compiler, so we fall back to a function-local statck thread_local (whew) variable with a wrapping registration function that needs to be called at least once by any thread using the SmallBufferAllocator.

Additionally, fix the test so that if it is called with --gtest_repeat it will pass.

Differential Revision: D23640790

fbshipit-source-id: f87c6e42ee0d1093dfdcc3cccb36ab2f7f52ea62