Reorganize SmallBufferAllocator to group per-thread data together
Summary: Initial impetus was to get around TSAN issues on mac.  I have previously disabled TSAN for a subset of moodycamel::ConcurrentQueue, but a recent change was using a different tact for thread cleanup.  To allow for use of enqueue_bulk with producer token, we need to avoid any thread shutdown order fiasco.  To do it, I grouped all the thread-local data that is required for cleanup together.  This winds up being cleaner and easier to follow anyway, so win-win.

Differential Revision: D23918416

fbshipit-source-id: f2e3cc8410dc9919f9876339bb362f17edf39cf1