Fix bug in thread_pool_test
Summary: I was inappropriately using the incorrect memory order for a load.  This meant that the compiler could migrate code below it to above it, and then there are bad race conditions.  Fixed to use correct memory order (note that for x86 at least, this doesn't change the instruction(s) used for the load, but just prevents the compiler from reordering code from below the load to above the load).

Differential Revision: D23995998

fbshipit-source-id: 19f147df79e9693d1196adb98ff08d546b53c59c