Make building dispenso benchmarks easier when building with cmake
Summary: Greatly simplify the building of dispenso benchmarks when using the cmake build.  This involved making some deps optional, and turning of some respective benchmarks.  For example, if TBB cannot be found, the benchmarks are built without TBB.  OMP and folly too.  Folly moreover has an additional flag that allows the user to disable folly benchmarking.  In actuality that flag is set for disabled by default because folly always seems to cause so many issues (case in point, I get crashes in folly code when I have it enabled with the RPM-installed folly on Fedora 33, due to internal checks in atomic::compare_exchange_strong where the wrong memory orders have been passed (namely that the failure memory order is stronger than the success memory order).

Reviewed By: darrelpalke

Differential Revision: D25771459

fbshipit-source-id: 9a9f85d96d783e62248b665099897d062e5c81a3