Change "then chain" implementation to avoid stack overflow potential
Summary:
Previously the implementation for Future::then was as a lock-free linked list chain of callables under the hood.  The chain was invoked by invoking the head, then the head would invoke next, it would invoke next and so on until the list was exhausted.

If `Future::then` is called enough times on a given Future, this could lead to stack overflow.  This in fact occurred at times in the Futures test on windows platforms (the test makes a future that waits 1ms to complete, and while it is not yet complete, we loop and call `then` on the future, resulting in possibly thousands of then continuations).

This changes the implementation from a chaining of function calls to a loop over the chain to invoke the continuations.  This should mean that we are bounded only by heap memory instead of stack in this situation, thereby resolving the potential stack overflow issue.

Reviewed By: darrelpalke

Differential Revision: D26160108

fbshipit-source-id: 1bb146a4ff3750f2374ecf152c379a694bd9f0ad