Enable a different polling period to be set for dispenso ThreadPool depending on platform
Summary: Some lower-power platforms might have a long/expensive context switch time relative to any length of sleep.  Dispenso has mostly been tested on desktop class platforms, and it can be desirable to reduce polling overheads, especially for idle cases.  This diff (a) modifies the default poll period from 100us to 200us.  It also (b) enables keying off of target platform in the build file to set a preprocessor macro to override the poll period.  The current diff keeps the default except on Android, where it doubles the poll period (for a 4x longer poll period than we had prior to this diff).

Reviewed By: thyu

Differential Revision: D26786277

fbshipit-source-id: de922c75866b4583e2219fbec18a2e5a4de838a0