Take two: try again to make MSVC happy with const lambda captures
Summary:
This time we use "=" default capture, which should slurp in all the values needed on MSVC.
See https://developercommunity.visualstudio.com/t/problems-with-capturing-constexpr-in-lambda/367326.  Fixed on newer compilers with latest C++ standard flag, but we don't have that version of msvc, and we want this code to be compatible with older compilers anyway.

Reviewed By: EscapeZero

Differential Revision: D27771405

fbshipit-source-id: 142513c0ede688c9fc4a49ffce36abcedefe97e7