Consolidate dispenso libary into one target (part 1)
Summary:
This diff just provides consolidated dispenso targets, one that can be used generally, and the other is local and only (visibility is empty) for testing programs built sans exceptions.

For now, existing external targets are maintained, but dispenso's tests and benchmarks build against the new consolidated target(s).

This change does not yet tackle moving other targets to depend on the consolidated target because targets exist in fbsource and also in ovrsource (outside of //arvr), so there will be a couple of follow up diffs to switch those over in fbcode and ovrsource separately.

Reviewed By: darrelpalke

Differential Revision: D27912526

fbshipit-source-id: 2c8397bedc1a6719995c575672d224cc7e3e2bd2