Remove Future::then()&& to fix race condition
Summary:
Ran future_benchmark for other work I was doing.  It crashed (appears to be a very rare crash).  Build and run with TSAN lead to lots of errors being spewed.  After a few days of debugging, I tracked this down to a race condition that covers a few lines of code.  The gist is that with then()&&, we were basically avoiding an atomic reference increment and decrement, but that increment was keeping the future guts alive until those few lines of code could finish.

I could have kept then()&& and just manually inc/dec the ref count, but this appears to be no faster than just using the regular then() function (I tried this and benchmarked).  Since there is no performance benefit, it is much better to simplify the code by removing then()&& and let it fall back to then().

Reviewed By: elliotsegal-fb

Differential Revision: D28944429

fbshipit-source-id: 7ab67884bf85c7daa7d5b034092b708d1755b6c8