Ensure that parallel loops wait if they are told to do so.
Summary:
It was possible before in a handful of cases to ignore the directive to wait on the task set prior to returning from parallel_for or for_each.  This diff fixes those cases.

Additionally, there was a case of divide-by-zero when the number of items passed into for_each was 0.  There is a fix for that too.

Differential Revision: D29005843

fbshipit-source-id: f29b80ba2b5dc29506c840ddb0c8eeaa6712264f