Fix bug in pipeline where wait wasn't being called properly prior to exiting the function
Summary: Found that the pipeline test was failing maybe 3 to 5 times out of 100.  After digging very deep, I found that wait wasn't working properly.  Visual inspection led me to believe that it should be correct.  But then, I found that wait wasn't even being called.  So yeah, duh that it was broken.  It's only just amazing that tests passed with such a high rate!

Reviewed By: mp3guy

Differential Revision: D29150722

fbshipit-source-id: e65ef31a035bb6810efd80de7b549fbccaf2cf6f