Modify the small_buffer_benchmark to remove repeated alloc of too few buffers
Summary: Some of the cases previously resulted in repeated allocation/deallocation of very small number of buffers.  This leads to a case where the thread essentially only goes to central store once at the beginning, and is very artificial.  Increase the number of allocations to force spill and fetch from central store.  This will be more appropriate for benchmarking the changes of the internal parameters of the small buffer allocator (since we won't hit huge cliffs as much).

Reviewed By: elliotsegal-fb

Differential Revision: D30119362

fbshipit-source-id: 21d46b324cf2b2e3493384aa09f65b5237a7b691