Use Schwarz counters to avoid untimely destruction order issues
Summary:
On Windows we had issues with the small buffer allocator globals going out of scope before the global thread pool, which lead to some destruction time crashes (guess we just got lucky on Linux/Mac).

While we could group the dispenso globals together in a specific order within a single translation unit, this would not solve the issue of other code using this from a global, so we employ Schwarz counters to fix the issue.

Reviewed By: elliotsegal-fb

Differential Revision: D30119361

fbshipit-source-id: d1d28cb8bbf50183f961cf62a81b5832d2a30aac