Fix dispenso small_buffer_allocator global destruction
Summary: Somehow it was still possible for ~SchwarzSmallBufferInit to be called after the reference count wound back to zero.  Still not quite sure why that occurs tbh.  But calling these globals destructors isn't really necessary because there will be no side effects by just letting the OS reclaim the memory at program shutdown, so the easy fix is to simply omit "proper" destruction, and it is still perfectly correct.

Reviewed By: YLouWashU

Differential Revision: D30240966

fbshipit-source-id: 418bf2d66bfe1691ca428bb9245c8d2fc6e26e03