Vastly reduce the minimum overheads of dispenso's small buffer allocator
Summary:
Reduce the minimum memory overhead of dispenso's small buffer allocator by a factor of 64.  This results in
a stiff increase in allocation cost, but it is still roughly an order of magnitude faster than new/delete backed by glibc malloc, and even if tcmalloc is used, small buffer allocator is still several times faster.

A factor of 64 means that if an embedded application was using 2 MB from small buffer allocator previously, they would now be using 32 KB, potentially making it's use more attractive.

I ran the future_benchmark as well to get a sense of impact on something doing a lot more than allocation, and the differences are in the noise.

Reviewed By: elliotsegal-fb

Differential Revision: D30119360

fbshipit-source-id: e56495ced841e39dec9d862e26d0a865041f6a63