Add a pretty fast thread-safe slabbed allocator
Summary: Implement a slabbed allocator suitable for any backing allocation that can slice (e.g. we can use ::malloc and ::free or cudaMalloc and cudaFree).  It is typically an order of magnitude faster than new/delete (shown in the benchmark), and should be much better for e.g. cudaMalloc/cudaFree

Reviewed By: yufangqq

Differential Revision: D32415194

fbshipit-source-id: 0c5989bd2948389c247a4b4f9d89150b5d22f8bd