ConcurrentVector: Fix bad end-index logic, which enabled double allocation through race condition
Summary:
Found ASAN issue with ConcurrentVector... need to investigate why ASAN tests aren't being run in CI.  There were two issues.

The first is a race condition for allocating underlying buffers.  This turned out to be due to allowing a `>=` when trying to decide whether to allocate one last buffer, when it should be `>`.  That is because another thread could be trying to allocate on the `==` concurrently.  Interestingly this didn't result in user data being thrown away because it was allocating the following, not current buffer, and this seems to be long enough in future to avoid writing to the thrown away buffer, even for TSAN builds (doesn't mean it couldn't happen, just doesn't).

The second issue is we were not deallocating enough in `clear`, again due to an off-by-one.  This could result in failure to free user data.

Reviewed By: yufangqq

Differential Revision: D32707302

fbshipit-source-id: 2e370774ad0a69c8c5b146d2a82fcaefd5c52367