Add Container type definitions to dispenso::ConcurrentVector
Summary:
The Container concept in C++ expects that a certain set of type aliases is defined on a class. To be able to use a dispenso::ConcurrentVector in a context that expects a Container, we should provide these type aliases.

This change adds the missing type aliases to ConcurrentVector.

reference and const_reference were already available under the names reference_type and const_reference_type. These were not touched so existing code will not break.

The requirements of the Container concept are described here: https://en.cppreference.com/w/cpp/named_req/Container

Note: The requirements on methods and operators were not checked.

Reviewed By: graphicsMan

Differential Revision: D33767315

fbshipit-source-id: 1389411fa4f4442dcd774183b7d0bc84cd246b85