Speed up Futures/CompletionEvent
Summary:
Mac Futures were slow.  Linux and Windows futures are much faster than Folly, but on Mac were about 2x
slower.  I was previously using an "optimization" for MacOS using fancy semaphore functionality.  As usual,
optimizing without performance testing wasn't a great idea.  Turns out that the fallback option is much faster, as
the overhead of creating and destroying semaphores is very high.  There may still be room to optimize this
further, either based on semaphores with pooling, or some other low-level mach mechanism, but this is already much
faster, and easy enough that we should just use this until there is time to invest on a lower overhead mechanism.

Reviewed By: elliotsegal-fb

Differential Revision: D37411519

fbshipit-source-id: 70883ae77b876db5ffa004f617a99b398785cd7f