Create a benchmark that can replicate pain for infrequent parallel_for use
Summary: I've seen multiple complaints that (a) dispenso causes too many context switches when idle and (2) on some platforms we see that dispenso is much slower because of the sleep/poll mechanism (Windows mostly).  I want to fix both of these.  This benchmark is an attempt to replicate the latter issue.

Reviewed By: EscapeZero

Differential Revision: D37940431

fbshipit-source-id: 3cee9786af70fb32cda24ba83c5343ea516e4b32