Use one less than desired number of total threads in thread pool
Summary:
Because dispenso aggressively uses the calling thread for computation in many scenarios (blocking parallel for, TaskSet wait, Future get), it makes sense to default to one fewer thread in the pool to avoid unnecessary context switching and contention.  This diff sets the default global thread pool size to number of hardware threads minus one.  It also uses one fewer thread for threaded benchmarks.

This also means we should adjust behavior for parallel_for and for_each to incorporate the calling thread differently.  It was already used before, but we need to adjust for the expectation that the total is allowed to incorporate the thread plus all pool threads.

Also include changes for ThreadPool to not hang if there are no threads in the pool.

Also include a fix for pipeline in the case the pool has no threads in it.

Reviewed By: EscapeZero

Differential Revision: D38211469

fbshipit-source-id: 83b322dba825df196a9cb73358b24b0b257b7c00