Add info for cost of actual loop code in latency benchmark
Summary: It is possible to pause and resume the timer in google benchmark, but then those times aren't used to tell the benchmark when to stop, leading to very very long runtimes.  This custom counter mechanism allows us to provide our own timings, so we can actually account for the cost of the parts we are interested in.  That is important for the latency benchmark because as computation becomes faster, the running time is dominated by sleep time.

Reviewed By: EscapeZero

Differential Revision: D38328775

fbshipit-source-id: 5dc4be9b316fe42222bfdfb586f09b03f21142ef