Eliminate race condition from benchmark
Summary: This benchmark had a subtle-ish race because the 0th thread would retain the 0th slot, while the rest of the threads would go up and down and may grab the 0th slot again (hence more than one thread could write to the 0th slot).  This was fixed by giving the 0th thread a dedicated slot.

Reviewed By: EscapeZero

Differential Revision: D38211471

fbshipit-source-id: 753c5b1df8cdc775701127c2063cd02eef5f7843