Implement an EpochWaiter for a lightweight signaling ThreadPool option
Summary: Implement a mechanism to wake idle threads with greater precision.  This addresses a critical issue on Windows where sleep wakeup time is treated as a hint at best (a 1ms sleep could last 10ms or more in some cases).  In addition, make this optional only.  While this yields better performance on Windows for nearly every workload, it is still a slight degredation in throughput-oriented tasks on Linux.  By making this optional, we can achieve the best performance on both platforms.  Additionally, we could look to increase the wait time to e.g. 30ms in the future combined with this feature to work better for infrequent work patterns on low-power devices.  This last bit is a TODO (still need to identify how to run these workloads jteller, muzoakbay).

Reviewed By: EscapeZero, FallenShard

Differential Revision: D38767177

fbshipit-source-id: bc59f8005eaa61961a712ddc1e33ea8f33e05399