Fix rare race when thread pool destructor is called concurrent with a TaskSet::wait for TaskSet still using the pool
Summary:
Pretty rare/complicated situation that would typically not occur because most people aren't going to use thread pool destruction as part of synchronization.  Even then, it would not be common to encounter this bug.  simple_pool_benchmark's BM_dispenso2 case triggers this on rare occasion, and I'd prefer to deem the usage valid, so this diff is a fix for that issue.  The fix is simply to allow the TaskSet to work steal off of the thread pool for all work rather than just the ones issued through that task set.

The scenario:  The main thread adds many tasks to the thread pool.  Those tasks launch new tasks through their task sets.

This was likely already a race before the up-stack changes, but those changes seem to have tickled things to become much more repeatable.

Reviewed By: mp3guy

Differential Revision: D38870362

fbshipit-source-id: af03c9674396012fd2abab091eb54d07d714880f