Refactor shared aspects of TaskSet and ConcurrentTaskSet into a base class
Summary:
As title.  Basically there is a bunch of shared info and functionality that is best maintained together.  This may make it possible to add further functionality in a simpler way (e.g. cancellation functionality).

Also, this refactor has already borne fruit, as it led me to what was logically a bug where we prematurely used ForcedQueueing scheduling.  After fixing this, we are now on par with TBB's performance in the pipeline benchmark.  This lead to a slight slowdown for the futures benchmark "BM_dispenso_taskset_tree", which was addressed by tuning the load parameter for the ConcurrentTaskSet (now slightly faster or slower depending on the case), and by adding a poweruser knob that allows avoiding an extra check for high load in the thread pool.

Reviewed By: EscapeZero

Differential Revision: D39905034

fbshipit-source-id: feedf033fe8ee242c05d67979fb17b757508f3bd