RFC Make cascading cancel of TaskSets opt-in
Summary: The new cascading cancel implementaiton was reasonably efficient, but did lead to performance regression.  This allows the feature to exist, but requires passing a parameter to (Concurrent)TaskSet constructor to say whether it will register with parent task sets to get cancellation notifications.  This allows folks already using the library to continue doing so without regression, but enables new use cases with cascading cancel.

Reviewed By: EscapeZero

Differential Revision: D39910024

fbshipit-source-id: 250a32a43ac3ee854291596cfe7507a3fb51f481