Fix minor bug in pool allocator
Summary: Previously we spun in a loop using fetch_add to spin lock.  Since we're using a 32-bit value, it is possible (even though extremely unlikely) to roll over back to zero, which would erroneously take the lock.  This hasn't actually occured in practice, but was noticed through visual inspection.  The fix is to use fetch_or which will ensure the value stays non-zero until the lock is released.

Reviewed By: elliotsegal-fb

Differential Revision:
D40869293

Privacy Context Container: L1124100

fbshipit-source-id: 54f1cd2ab8c58823807dc798f6e0e3638a84de2f