Attempt to make cpu-counter based timing more accurate
Summary: Accuracy of counter-based timing had an error of about 1 part in 5000, which isn't super bad, but isn't super great.  With these changes, we achieve error of about 1 in 100K.  Additionally, previously the first call to getTime() would take 10ms.  I've changed this to happen before main(). The downside is that startup will now take 50ms; I think 1 20th of a second is reasonable to gain fast and accurate timer capability.  For Mac this can be avoided, since we can query the system for the TSC frequency.

Reviewed By: EscapeZero

Differential Revision:
D41775291

Privacy Context Container: L1124100

fbshipit-source-id: 39a5c21f1698b917bfe6ab4841f2f5b61c117d5a