Ensure that small buffer allocator cannot try to return resources erroneously after main exits
Summary: There is a niche situation where threads that use the small buffer allocator exit after the underlying central store queue is already destroyed.  When that happens, those threads' thread_local storage would be destroyed, and try to return the resources to the central store, which is no longer valid.  This diff attempts to guard against returning resources when the central store is invalid by making the schwarz counter do double duty as a lock.

Reviewed By: elliotsegal-fb

Differential Revision:
D42019805

Privacy Context Container: L1124100

fbshipit-source-id: 146ded46ec3320d04bc4b0b9c615bd18f7ff8afe