Ensure that moodycamel ProducerToken is released from dispenso's code
Summary: Previously it could be possible that the ProducerToken could be released from the dispenso dll even though it might be allocated in a different dll in the case that a TaskSet was used.  This ensures that the allocation happens in the same place deallocation happens, which matters for Windows + dlls.

Reviewed By: EscapeZero

Differential Revision: D44436541

fbshipit-source-id: 32f2d9549dbbdbd911acdbc969d4c9fb858bf351