Enable limiting chunk sizes to parallel_for
Summary: Very often the input workloads for a parallel_for are variable.  In cases where each loop iteration is lightweight, parallel_for had high overhead when there wasn't much work to do.  We add a new option `minItemsPerChunk` which controls the sizes of chunks that are divvied for threads to work on when load balancing is kAuto or kStatic.  This option will be ignored if a concrete chunk size is provided to makeChunkedRange.

Reviewed By: EscapeZero

Differential Revision: D49334866

fbshipit-source-id: 62040329e0a6dcf872bc8a27514b821403097ffe