Work around TSAN "bug"
Summary:
More recent versions of TSAN appear to be causing sleeps to happen regularly on atomic operations.  For some lock-free algorithms, this churn makes runtime extremely slow, causing tests to time out (though these do make slow progress, and if allowed to run, will complete without TSAN errors).  One of the main areas this happens is when multiple threads are growing regions of a ConcurrentVector.  One thread, and only one thread acquires any region of the vector to allocate.  While that region is being allocated any other thread requiring that region stalls in a loop until the newly allocated region is published.  The single publisher stores the pointer atomically.  I've found experimentally that TSAN likes to wait on some kind of semaphore inside this store, prior to the actual atomic operation happening.  When all other threads are waiting for this to be published, it seems that this semaphore may not be actively signaled.  However, if one or more of the waiting threads give up their time slice to the OS, TSAN quickly signals the semaphore, allowing the buffer to be published.

To be clear, there is no indication that there is a bug in dispenso, and really this is more of an indicator that TSAN has an issue.  However, we'd still like to be able to utilize TSAN for all of its magic goodness in many other surrounding scenarios, so we employ this workaround here.

Reviewed By: elliotsegal-fb

Differential Revision: D50202176

fbshipit-source-id: ddd1f1e843a81dc3d3f70426175ac8c1734459d4