Control thread pool scheduling parameters via compile-time macro definitions
Summary:
Dispenso defaults to simple low-ish latency polling on Linux, and slightly more complicated notification-based scheduling on Windows.  If pretty much all computation threads in the system are using dispenso, this is ideal.  However, if heavy computation and inter-thread handoff are occuring outside of dispenso, allowing low-latency polling for nearly hardware-concurrency threads can have dramatic downsides.  It is recommended in those cases to reduce the size of the global thread pool, and additionally to change to the wakeup-based scheduling mechanism with higher-latency polling to reduce context switch churn that can affect external (e.g. condition_variable) thread handoffs.

Note that this change does *not* modify any of the defaults, though it might be worth consideration in the future.

Reviewed By: elliotsegal-fb

Differential Revision: D50331601

fbshipit-source-id: 703997edee028bfd5ee7f7f4b8cd36bf9892e410