Fix shutdown "fix" for tracking NewThreadInvoker threads
Summary:
Previously, I tried to fix the wait for detached threads in NewThreadInvoker by using atexit.  While this likely helps, it wasn't really the solution.  The main problem was that in `remove`, I was calling `notify(0)`, which modifies the CompletionEventImpl's internal atomic.  That is fine for cases where no one will ever modify the atomic after notify, but in this case, that guarantee didn't hold, creating a race condition.

I was able to fix this via a `tryNotify` implementation, which works with the FUTEX implementation, the WakebyAddress/WaitOnAddress, and the MACH semaphore solutions, but I was unable to come up with a performant solution for the fallback.

In this case, we won't expect a high throughput of NewThreadInvoker schedules, and in any case, the cost of starting a thread dwarfs the cost of tracking the threads, so for clean/readable code, I just implemented this with mutex+condition_variable, which avoids the race condition.

Reviewed By: RomanFedotovFB

Differential Revision: D51043895

fbshipit-source-id: f7b8337a81f282d0f685821da8cbd274a6267b4c