Fix small buffer allocation for 4-byte and smaller allocations
Summary: In 32-bit builds, we can commonly have requests for 4-byte structures (basically a pointer).  Expand dispenso to handle that case.  Additionally, take care to ensure that the smallest allocation size (4 bytes) will be used for requests of 1 or 2 bytes even though that will not be something dispenso itself does (because perhaps a user will want to use it directly for small allocations).

Reviewed By: Dvad

Differential Revision: D54136829

fbshipit-source-id: 1dca1212e1e8e05232a78be387a31233efbea63b