Fix use of folly thread pool in benchmark
Summary: Previously we were passing the number of threads as the priority.  This meant that when we have a lot of threads, this could overflow and become e.g. a negative value, which was an invalid priority.  We don't actually want to muck with priority anyway, so use of that particular constructor was a mistake.  Now we just provide the number of threads.  I *think* this could be something that Folly changed silently at some point, but it could have been an error from the start.

Reviewed By: RomanFedotovFB

Differential Revision:
D58752369

Privacy Context Container: L1263768

fbshipit-source-id: 024d687d03092f3be968d6c600031aa68c3272c0